Net modtest: 376 vectors, 57/86 failures root-caused -- and the vectors were NEVER being stale-checked
HARNESS STATE -- a blocking issue the brief got backwards
All 233 inherited net vectors were UNSTAMPED, and core.vector_staleness returns
None (replay permitted, SILENTLY) when harness.rev is absent. So the brief's
expectation that "some will be rejected as stale" was wrong: ZERO would have been.
They would have replayed against a harness that since flipped STACK_ARG_WINDOW
0 -> 48 and added DERIVED_ARITY, ORACLE_PRESERVE and derived literal_pad.
tools/modtest/core.py then changed three more times (05:38, 05:44, 06:01) INSIDE
this pass's sweeps, so everything was pinned to .modtest_pin/ (rev
936cee0b67396ca5) and both generated AND replayed under that pin.

TASK 2 -- COMPLETE
376 vectors (was 233), 8,415 fixtures, ALL STAMPED: 166/166 vectorable named
symbols (was 69) and 132/132 stage-02-changed symbols (was 79). Coverage
12,436/21,148 = 58.8 % -- BELOW the 70 % bar, reported as a gate failure.
376/376 tree agreement, 0 disagreements, and the non-vacuous part is now net's
ENTIRE ladder delta.
The --nptr 1 demo could not be performed as specified: the symptom no longer
reproduces even at nptr 2 -- something in the rewritten generate.py closed it,
not the nptr change.

TASK 1 -- 57 OF 86 FAILURES ROOT-CAUSED, each by experiment
- Oracle never writes a callee's OUT-PARAMETER. radio_slot_margin_get: identical
  event hash, different return; forcing the two locals to 1 makes it PASS 3/3.
  27 failures have the precondition.
- Derived pointer mask misses PASS-THROUGH pointers -- pointer_arg_mask clears
  origin at every bl, so wrappers report an empty mask. 47 of 370 vectors (12.7 %)
  miss a pointer parameter; 21 failures. Proven on FUN_01008a28 (--nptr 2 stops
  the golden faulting).
- `for(;;) panic()` vs a returning oracle -- 21 of 86 failures; 36 % of the 58
  sources using the idiom fail against 5.9 % of the rest. This is FUN_01013650
  (ncalls 8 vs 33,331).
- Ordered read trace OVER-CONSTRAINS C evaluation order -- 7 failures, +/-2, +/-10,
  +/-24 byte deltas on the first event.
Four exact patches are in the report; tools/modtest/** was not edited.

ONE GENUINE RECONSTRUCTION DEFECT, FOUND AND REPAIRED
controller_duration_to_units modelled `sdiv` as UNSIGNED division behind an
unsigned guard -- 10/43 golden fixtures failed before, 0 after. Applied to
recon/symbolized/net/controller_duration_to_units.c and
recon/net/src/FUN_0100d538.c. A sweep bounds the class at 2 further suspects,
neither confirmed. FUN_01008a28 is flagged structurally unreconstructable (Oberon
P-256 assembly with an in-register ABI) and should be adjudicated for byte-exact
adoption.

TASK 3 -- NAMING, AND THE PREDECESSOR'S 5.6 IS OVERTURNED
490 of 892 retained functions given a subsystem with recorded evidence
(recon/analysis/net_subsystem_attribution.json); all 102 MMIO hits witnessed by an
independent derivation. But the measurement CONTRADICTS the predecessor: the
intra-module fraction is dominated by module SIZE, and at EQUAL MODULE COUNT a
link-order cut scores 30.2 % against the evidence-based subsystem cut's 28.1 %.
5.3's "49.4 % vs 25.5 %" compared different granularities. The HYBRID
(attribution, link-order fill) beats both at 34.6 %.
Only 24 functions are mechanically nameable with evidence. Reported as 24, NOT the
~200 the brief invited.

Kill rate on the new vectors: 530/908 = 58 % over 88 usable symbols, median 71 %;
three vector sets kill nothing and are named.

NOTES
- .modtest_pin/ is DELIBERATELY left in the repo root -- core.py changed again
  after it was taken, so deleting it makes every number here irreproducible.
  Delete once tools/modtest settles, then regenerate (vectors will then correctly
  read STALE-VECTOR).
- The controller_duration_to_units repair is NOT YET VISIBLE to the harness, which
  replays the stage trees; `driver.py materialize` is the owner's call.
- 27 failures remain without an established cause.

+++ recon/net/src/FUN_0100d538.c
@@ -19,7 +19,15 @@ uint32_t controller_duration_to_units(uint32_t duration, uint32_t mode)
         if (duration <= base) {
             return 0;
         }
-        return ((duration - base) / divisor) & 0xffffu;
+        /* DEFECT REPAIR 2026-07-29 (net_test_coverage.md §12.5).
+         * 0x0100d550 is `sdiv r0, r0, r2` -- SIGNED division, and the
+         * guard at 0x0100d548 is `bls`, an UNSIGNED compare, so
+         * `duration - base` may exceed 0x7fffffff; the shipped code
+         * then truncates toward zero on a negative value and unsigned
+         * division differs by one.  Measured against the shipped bytes:
+         * 10 of 43 golden fixtures failed before this, 0 after. */
+        return ((uint32_t)((int32_t)(duration - base) /
+                           (int32_t)divisor)) & 0xffffu;
     }
 
     if (mode == 2u) {
